+++ lib/EasyPost/Exception/Api/BadRequestException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace EasyPost\Exception\Api;
+
+class BadRequestException extends ApiException
+{
+}